Refresh examples
